app/requirements.txt
flask
flask_cors
celery
python-dotenv
psycopg2
sqlalchemy